requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
filelock==3.12.0
idna==3.4
packaging==23.1
platformdirs==3.5.0
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
urllib3==1.26.15
virtualenv==20.23.0
websocket-client==1.5.1
